requirements.txt
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
xgboost>=1.7.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
joblib>=1.3.0
jieba>=0.42.0
nltk>=3.8.0
requests>=2.31.0